requirements.txt
discord
psycopg2
python-dotenv
asyncio